requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.51
botocore==1.35.51
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.5
langchain-aws==0.2.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-text-splitters==0.3.1
langgraph==0.2.39
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.137
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.10
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
s3transfer==0.10.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
wheel==0.44.0
yarl==1.17.0
